Move siege events
