+++ src/main/java/net/countercraft/movecraft/warfare/features/siege/events/SiegeEvent.java
@@ -1,18 +1,20 @@
-package net.countercraft.movecraft.warfare.events;
+package net.countercraft.movecraft.warfare.features.siege.events;
 
 import org.jetbrains.annotations.NotNull;
+
+import net.countercraft.movecraft.warfare.features.siege.Siege;
+
 import org.bukkit.event.Event;
-import net.countercraft.movecraft.warfare.siege.Siege;
 
 public abstract class SiegeEvent extends Event {
-    @NotNull protected final Siege siege;
+    @NotNull private final Siege siege;
 
     public SiegeEvent(@NotNull Siege siege) {
         this.siege = siege;
     }
 
     @NotNull
-    public final Siege getSiege() {
+    public Siege getSiege() {
         return siege;
     }
 }
